requirements.txt
Django==2.0.6
django-cors-headers==2.4.0
djangorestframework==3.8.2
future==0.16.0
junitparser==1.2.2
pytz==2018.4
drf-extensions==0.4.0